+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## 0.4.11
 
 * feat: export strCompare from './utils';
+* feat: new `props.smoothScroll` default: chrome < 85 ? true : false.
 * optimize: if not virtual table, need not do auto resize function.
 * fix: `StkTable.setHighlightDimCell` first parameter type accept number.
-* fix: `insertToOrderedArray` emptyToBottom bugfix
+* fix: `insertToOrderedArray` emptyToBottom bugfix.
 
 ## 0.4.10